Move code to `features/view-as-user`
