+++ web-admin/src/features/view-as-user/ViewAsUserPopover.svelte
@@ -8,7 +8,7 @@
   import { matchSorter } from "match-sorter";
   import { createEventDispatcher } from "svelte";
   import { createAdminServiceSearchProjectUsers, V1User } from "../../client";
-  import { errorStore } from "../errors/error-store";
+  import { errorStore } from "../../components/errors/error-store";
   import { viewAsUserStore } from "./viewAsUserStore";
 
   export let organization: string;
